% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brown.R
\name{Brown}
\alias{Brown}
\title{Brownian Motion.}
\usage{
Brown(t, n)
}
\arguments{
\item{t}{the time.}

\item{n}{the number of paces of Brownian Motion.}
}
\value{
plot figure\code{n}
}
\description{
plot of st Brownian Motion，before use the function,you should use package e1071.
}
\examples{
\dontrun{
Brown(1,6)
}
}
