% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sseR.R
\name{sseR}
\alias{sseR}
\title{Use stratified sampling to estimate the integration.}
\usage{
sseR(n)
}
\arguments{
\item{n}{the power of x.}
}
\value{
estimate value \code{n}
}
\description{
Estimate the integration of x^nexp(-x) from 0 to 1.
}
\examples{
\dontrun{
re1R = sseR(5)
}
}
