% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xtC}
\alias{xtC}
\title{etimate possion process.}
\usage{
xtC(N, lambda)
}
\arguments{
\item{lambda}{the parameter of the exponential distribution.}
}
\value{
a result \code{n}
}
\description{
we can get in in R.
}
\examples{
\dontrun{
xtC(100,1)
}
}
