% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtR.R
\name{xtR}
\alias{xtR}
\title{Creat a possion process}
\usage{
xtR(N, lambda)
}
\arguments{
\item{N}{the number of Tn.}

\item{lambda}{the parameter of the exponential distribution.}
}
\value{
a estimate result \code{n}
}
\description{
Estimate the number of the events when t=10.
}
\examples{
\dontrun{
re2R=numeric(1e4)
for(i in 1:1e4){
re2R[i] = xtR(100,1)}
round(mean(re2R))
}
}
